An I2C read now reads any register. Add code for an Icom AH-4 antenna tuner.

+++ n2adr_lib/tx_freq_to_band.c
@@ -7,7 +7,7 @@
 #include "../hl2ioboard.h"
 
 #define MAX_FREQ_INDEX	23
-int tx_freq_to_band(uint64_t freq)
+uint8_t tx_freq_to_band(uint64_t freq)
 {
 // Band  137k  500k  160  80  60  40  30  20  17  15  12  10   6   4   2  1.25  70cm  33cm  23cm  13cm  9cm  5cm  3cm
 // Index    1     2    3   4   5   6   7   8   9  10  11  12  13  14  15    16    17    18    19    20   21   22   23
